Allow rejoining the room with different peer
Currently we are ignoring join message from participant who already joined the room. This is causing issue in case where participant lost connectivity and needed to open new websocket (ip address change for example). In this PR we will kick out old participant in that case,
close the transport and use new peer to relay messages to.